Addressing PR comments
Refactor Spec classes into separate files
Fix errors in sample generation
