+++ src/main/java/com/google/api/codegen/php/PhpGapicContext.java
@@ -16,6 +16,9 @@ package com.google.api.codegen.php;
 
 import com.google.api.codegen.ApiConfig;
 import com.google.api.codegen.GapicContext;
+import com.google.api.codegen.metacode.FieldSetting;
+import com.google.api.codegen.metacode.InitCode;
+import com.google.api.codegen.metacode.InitCodeLine;
 import com.google.api.tools.framework.model.Interface;
 import com.google.api.tools.framework.model.Method;
 import com.google.api.tools.framework.model.Model;
@@ -255,6 +258,20 @@ public class PhpGapicContext extends GapicContext implements PhpContext {
         // Types that do not need to be modified (e.g. TYPE_INT32) are handled here
         return value;
     }
+  }
+
+  /**
+   * Determine whether a given InitCodeLine is a method parameter. This is important for map fields
+   * because flattened map parameters accept an associative array, while map fields on a request
+   * object must be a sequential array of MapEntry objects.
+   */
+  public boolean initLineIsParameter(InitCode initCode, InitCodeLine initCodeLine) {
+    for (FieldSetting fieldSetting : initCode.getArgFields()) {
+      if (initCodeLine.getIdentifier().equals(fieldSetting.getIdentifier())) {
+        return true;
+      }
+    }
+    return false;
   }
 
   // Constants
