+++ src/main/resources/com/google/api/codegen/py/method_sample.snip
@@ -91,9 +91,9 @@
 
 # Helper method for initLineMap()
 @private initMap(line)
-  @join key : line.getElementIdentifierKeys vertical
+  @join key : line.getElementIdentifierKeys vertical on ","
     @let identifierValue = line.getElementIdentifierValue(key)
-      {@context.renderPrimitiveValue(line.getKeyType, key)}, {@identifierValue},
+      {@context.renderPrimitiveValue(line.getKeyType, key)}: {@identifierValue}
     @end
   @end
 @end
