+++ src/test/java/com/google/api/codegen/testdata/java_main_library.baseline
@@ -60,7 +60,6 @@ import com.google.example.library.v1.PublishSeriesResponse;
 import com.google.example.library.v1.Shelf;
 import com.google.example.library.v1.SomeMessage;
 import com.google.example.library.v1.UpdateBookIndexRequest;
-import com.google.example.library.v1.UpdateBookIndexRequest.IndexMapEntry;
 import com.google.example.library.v1.UpdateBookRequest;
 import com.google.protobuf.ByteString;
 import com.google.protobuf.Empty;
@@ -1822,7 +1821,7 @@ public class LibraryServiceApi implements AutoCloseable {
    *   String formattedName = LibraryServiceApi.formatBookName("[SHELF]", "[BOOK]");
    *   String indexName = "default index";
    *   String indexMapItem = "";
-   *   Map&lt;String, UpdateBookIndexRequest.IndexMapEntry&gt; indexMap = new HashMap&lt;&gt;();
+   *   Map&lt;String, String&gt; indexMap = new HashMap&lt;&gt;();
    *   indexMap.put("default_key", indexMapItem);
    *   libraryServiceApi.updateBookIndex(formattedName, indexName, indexMap);
    * }
@@ -1859,7 +1858,7 @@ public class LibraryServiceApi implements AutoCloseable {
    *   String formattedName = LibraryServiceApi.formatBookName("[SHELF]", "[BOOK]");
    *   String indexName = "default index";
    *   String indexMapItem = "";
-   *   Map&lt;String, UpdateBookIndexRequest.IndexMapEntry&gt; indexMap = new HashMap&lt;&gt;();
+   *   Map&lt;String, String&gt; indexMap = new HashMap&lt;&gt;();
    *   indexMap.put("default_key", indexMapItem);
    *   UpdateBookIndexRequest request = UpdateBookIndexRequest.newBuilder()
    *     .setName(formattedName)
@@ -1890,7 +1889,7 @@ public class LibraryServiceApi implements AutoCloseable {
    *   String formattedName = LibraryServiceApi.formatBookName("[SHELF]", "[BOOK]");
    *   String indexName = "default index";
    *   String indexMapItem = "";
-   *   Map&lt;String, UpdateBookIndexRequest.IndexMapEntry&gt; indexMap = new HashMap&lt;&gt;();
+   *   Map&lt;String, String&gt; indexMap = new HashMap&lt;&gt;();
    *   indexMap.put("default_key", indexMapItem);
    *   UpdateBookIndexRequest request = UpdateBookIndexRequest.newBuilder()
    *     .setName(formattedName)
