+++ src/test/java/com/google/api/codegen/testdata/python_main_library.baseline
@@ -857,7 +857,7 @@ class LibraryServiceApi(object):
           >>> name = api.book_path('[SHELF]', '[BOOK]')
           >>> index_name = ''
           >>> index_map_item = ''
-          >>> index_map = {'default_key', index_map_item,}
+          >>> index_map = {'default_key': index_map_item}
           >>> api.update_book_index(name, index_name, index_map)
 
         Args:
